requirements.txt
fastapi
uvicorn
pydantic
python-multipart
Pillow
tensorflow==2.15.0
python-jose
passlib[bcrypt]
